requirements.txt
# ============================================================
# DEPENDENCIAS — compatibles con Python 3.13
# Sistema de Asignación Prioritaria de Internado
# ============================================================

# --- Core del backend ---
Django==4.2.8
djangorestframework==3.14.0

# --- Base de datos ---
psycopg2-binary==2.9.10

# --- Cache en producción ---
redis==5.2.1
django-redis==5.4.0

# --- Autenticación JWT ---
djangorestframework-simplejwt==5.4.0

# --- Documentación OpenAPI ---
drf-spectacular==0.27.0

# --- Variables de entorno ---
python-dotenv==1.0.1

# --- Filtros ---
django-filter==24.2

# --- CORS ---
django-cors-headers==4.3.1

# --- Algoritmo IA (compatibles con Python 3.13) ---
xgboost==2.1.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3